// frontend/src/api/models/createMemoryJob.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { JobParams } from './jobParams';

export interface CreateMemoryJob {
  endChapterNum?: number | null;
  memoryGroupId: string;
  params: JobParams;
  startChapterNum?: number | null;
}